Roadmap UI: Allow the UI to load all route configs up front instead of one at a time on demand. (#132)
* Load all route configs up front instead of as needed.

The proposed UI needs full route configs loaded up front.  This is so that the "spider" map can show all routes that stop near a point, and so that the "home" screen can list all routes and directions.  The route configs will be used together with the precomputed trip/wait isochrone data to show grades and rankings for all routes.

This change is meant to be easily reversible in case issues are discovered with it.  There are other, clunkier ways for the front end to load all the route configs (one call per route) but avoiding that for now.

* Added todo comment about refactoring get_route_config.

* Remove prettyprinting from /routes and /routeConfig to save bandwidth.