MapShield component introduction
Introduction of the MapShield component, part of the UI roadmap.  This component takes a full list of route configs and uses it to show routes that run near a point on a map.  The routes also include "shields" (circled route numbers) at the terminals. The plotted routes can be clicked but currently do not dispatch any events, as we will probably redo where state is stored in the roadmap UI.