Codeowners to auto suggest reviewers
This will help with auto suggesting reviewers

It be better if we moved the python stuff into a `/backend/` folder so we can maintain this cleaner as a mono repo and make the file structure of this repo suggest code owners properly. Ideally better than maintaining a document that suggest code owners to tag.