Try to clip the isochrone overlay to the service area
We use turf's geoJson processing tools to clip all the geoJson as it is
added to the layer. This is slow, and buggy when
malformed/self-intersecting polygons are passed (which they are).

This commit also stops you from clicking a second point to set an
endLatLng that's outside the service area (i.e. the sea).