Skip updating GTFS if no active routes on date being run (#627)
https://github.com/trynmaps/metrics-mvp/pull/627

Agencies normally update their GTFS a few days ahead so the one that gets downloaded ends up having no active routes. This could result in the wrong routes being shown on the dates just before the changes take effect, if for example a route is added or removed.

If there are no active routes on the provided date, save_routes will raise an exception and the routes will not be updated. The older version of routes will be used by the backend, which should result in the expected behaviour.