Use recentmost date key if date is not active in GTFS (#648)
As to identify the GTFS date corresponding to a given
date (ie. the GTFS date containing the scheduled trips
for that date) a date_keys lookup is done.

The lookup is done on the date_keys file, which persists
the GTFS date for each date that was processed, as to allow
for re-running pre-computation on older GTFS's.

Since the recentmost GTFS file Muni put out expired on 2020-05-29,
the pre-computation as well as the API failed due to the date key
lookup failing.

This PR modifies the get_date_key function as to return the
recentmost date key corresponding to the same day of the week.
Note that it does not take holidays into account. It raises
an exception if it cannot find a suitable date_key to use.

It prints a warning if it succeeds, although a future PR
could send out an email so that we are aware the GTFS is outdated.