Change Metrics API production endpoint (#658)
* Change Metrics API production endpoint

As the frontend is hosted on S3 the metrics API endpoint needs to be changed to point to the API running on google cloud, which can be reached through data.opentransit.city

* Use sf-muni instead of data

Due to metrics api only having certs for muni and sf-muni configured. A new one for data would have to be added.